Updates
- added script to train online RL agents on parallel environment
- updated config_changer.py to allow parallel_env
- fixed utility_logistic algorithm to use 0.5 if nan
- made first attempt at improving LSPI algorithm with float64 datatype instead of
float32

+++ NetworkAgent/fast_lspi/agent_random.py
@@ -65,7 +65,7 @@ class FastLSPI:
 
     def initialize_Q_weights(self) -> None:
         self.k = self.hidden_dims[-1] + self.num_actions
-        self.w_tilde = torch.randn((self.k, 1), device="cuda:0")
+        self.w_tilde = torch.randn((self.k, 1), dtype=torch.float64, device="cuda:0")
         self.full_rank_reached = False
 
     def store_to_buffer(
